requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
docker
python-dotenv
werkzeug
gunicorn
